cng: add support for windows/386
