+++ .github/workflows/test.yml
@@ -6,15 +6,25 @@ jobs:
       fail-fast: false
       matrix:
         go-version: [1.25, 1.26]
-        venv: [windows-2022, windows-2025, windows-11-arm]
+        target:
+          - venv: windows-2022
+            goarch: amd64
+          - venv: windows-2025
+            goarch: amd64
+          - venv: windows-11-arm
+            goarch: arm64
+          - venv: windows-2022
+            goarch: '386'
         fips: [1, 0]
-    runs-on: ${{ matrix.venv }}
+    runs-on: ${{ matrix.target.venv }}
+    env:
+      GOARCH: ${{ matrix.target.goarch }}
     steps:
     - uses: actions/setup-go@b7ad1dad31e06c5925ef5d2fc7ad053ef454303e # v7.0.0
       with:
         go-version: ${{ matrix.go-version }}
         # There's not a Mirosoft build of Go 1.25 for Windows 11 ARM, so use the stdlib instead.
-        go-download-base-url: ${{ (matrix.venv != 'windows-11-arm' || matrix.go-version != 1.25) && 'https://aka.ms/golang/release/latest' || '' }}
+        go-download-base-url: ${{ (matrix.target.venv != 'windows-11-arm' || matrix.go-version != 1.25) && 'https://aka.ms/golang/release/latest' || '' }}
     - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
 
     - name: Set FIPS mode
