+++ cng/aes_test.go
@@ -10,6 +10,8 @@ import (
 	"bytes"
 	"crypto/cipher"
 	"fmt"
+	"math"
+	"runtime"
 	"strings"
 	"testing"
 
@@ -208,13 +210,11 @@ func TestSealPanic(t *testing.T) {
 	assertPanic(t, func() {
 		gcm.Seal(nil, make([]byte, gcmStandardNonceSize-1), []byte{0x01, 0x02, 0x03}, nil)
 	})
-	assertPanic(t, func() {
-		// maxInt is implemented as math.MaxInt, but this constant
-		// is only available since go1.17.
-		// TODO: use math.MaxInt once go1.16 is no longer supported.
-		maxInt := int((^uint(0)) >> 1)
-		gcm.Seal(nil, make([]byte, gcmStandardNonceSize), make([]byte, maxInt), nil)
-	})
+	if runtime.GOARCH != "386" {
+		assertPanic(t, func() {
+			gcm.Seal(nil, make([]byte, gcmStandardNonceSize), make([]byte, math.MaxInt), nil)
+		})
+	}
 }
 
 func TestAESInvalidKeySize(t *testing.T) {
