+++ internal/bcrypt/authinfo_windows_386.go
@@ -1,0 +1,23 @@
+//go:build windows && 386
+// +build windows,386
+
+package bcrypt
+
+// https://docs.microsoft.com/en-us/windows/win32/api/bcrypt/ns-bcrypt-bcrypt_authenticated_cipher_mode_info
+type AUTHENTICATED_CIPHER_MODE_INFO struct {
+	Size           uint32
+	InfoVersion    uint32
+	Nonce          *byte
+	NonceSize      uint32
+	AuthData       *byte
+	AuthDataSize   uint32
+	Tag            *byte
+	TagSize        uint32
+	MacContext     *byte
+	MacContextSize uint32
+	AADSize        uint32
+	_              uint32
+	DataSize       uint64
+	Flags          uint32
+	_              uint32
+}
